requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.44.0
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-rename-app==0.1.3
django-webpack-loader==1.0.0
djangorestframework==3.12.4
google-api-core==1.31.0
google-auth==1.32.1
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dotenv==0.18.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
